# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Hotel API",
        "description": "API for hotel management",
        "version": "1.0"
    },
    "produces": [
        "application/json"
    ],
    "host": "127.0.0.1:3000",
    "basePath": "/",
    "paths": {
...
        "/clients": {
            "get": {
                "operationId": "clients",
                "tags": [
                    "Clients"
                ],
                "description": "Récupère la liste des clients",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "id_client": {
                                        "type": "integer"
                                    },
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "registration_date": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    }
}